+++ src/openms/source/ANALYSIS/ID/SimpleSearchEngineAlgorithm.cpp
@@ -795,9 +795,9 @@ void SimpleSearchEngineAlgorithm::postProcessHits_(const PeakMap& exp,
             ah.sequence = c;
             ah.peptide_mod_index = mod_pep_idx;
             ah.score = score;
-            ah.prefix_fraction = (double)detail.matched_b_ions/(double)c.size();
-            ah.suffix_fraction = (double)detail.matched_y_ions/(double)c.size();
-            ah.mean_error = detail.mean_error;
+            ah.prefix_fraction = static_cast<float>((double)detail.matched_b_ions / (double)c.size());
+            ah.suffix_fraction = static_cast<float>((double)detail.matched_y_ions / (double)c.size());
+            ah.mean_error = static_cast<float>(detail.mean_error);
             ah.matched_b_ions = static_cast<uint16_t>(detail.matched_b_ions);
             ah.matched_y_ions = static_cast<uint16_t>(detail.matched_y_ions);
             ah.isotope_error = low_it->second.second;
